# src/definitions/swagger.yaml
basePath: /
definitions:
    IUser:
        description: ""
        properties:
            id:
                type: string
                description: ""
            email:
                type: string
                description: ""
            firstname:
                type: string
                description: ""
            lastname:
                type: string
                description: ""
            username:
                type: string
                description: ""
            mobileNumber:
                type: string
                description: ""
            createdAt:
                type: string
                format: date-time
                description: ""
            updatedAt:
                type: string
                format: date-time
                description: ""
        type: object
        required:
            - id
            - email
            - firstname
            - lastname
            - username
            - mobileNumber
            - createdAt
            - updatedAt
...
    IUserPayment:
        description: ""
        properties:
            id:
                type: string
                description: ""
            userId:
                type: string
                description: ""
            bettingLeagueId:
                type: string
                description: ""
            paid:
                type: boolean
                description: ""
            displayed:
                type: boolean
                description: ""
        type: object
        required:
            - id
            - userId
            - bettingLeagueId
            - paid
            - displayed
    IUserRequest:
        description: ""
        properties:
            id:
                type: string
                description: ""
            userId:
                type: string
                description: ""
            bettingLeagueId:
                type: string
                description: ""
            decided:
                type: boolean
                description: ""
            accepted:
                type: boolean
                description: ""
        type: object
        required:
            - id
            - userId
            - bettingLeagueId
            - decided
            - accepted
    IUserSetting:
        description: ""
        properties:
            id:
                type: string
                description: ""
            userId:
                type: number
                format: double
                description: ""
            bettingLeagueId:
                type: number
                format: double
                description: ""
            emailBetNotification:
                type: boolean
                description: ""
            emailRankingNotification:
                type: boolean
                description: ""
        type: object
        required:
            - id
            - userId
            - bettingLeagueId
            - emailBetNotification
            - emailRankingNotification
    IUserSpecialBet:
        description: ""
        properties:
            id:
                type: string
                description: ""
            specialBetId:
                type: number
                format: double
                description: ""
            bettingLeagueId:
                type: number
                format: double
                description: ""
            userId:
                type: number
                format: double
                description: ""
            seriesHomeTeam:
                type: number
                format: double
                description: ""
            seriesAwayTeam:
                type: number
                format: double
                description: ""
            specialBet:
                type: number
                format: double
                description: ""
            dateTime:
                type: string
                format: date-time
                description: ""
        type: object
        required:
            - id
            - specialBetId
            - bettingLeagueId
            - userId
            - seriesHomeTeam
            - seriesAwayTeam
            - specialBet
            - dateTime
info:
    license:
        name: MIT
    title: tipapp-api
    version: 0.0.1
paths:
...
    '/api/users/{id}':
        get:
            operationId: UsersGetUser
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IUser'
            description: ""
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
        put:
            operationId: UsersUpdateUser
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IUser'
            description: ""
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
                -
                    description: ""
                    in: body
                    name: user
                    required: true
                    schema:
                        type: object
        delete:
            operationId: UsersDeleteUser
            produces:
                - text/html
            responses:
                '204':
                    description: 'No content'
            description: ""
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
...
    /api/userPayments:
        get:
            operationId: UserPaymentsGetUserPayments
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/IUserPayment'
            description: ""
            parameters: []
        post:
            operationId: UserPaymentsCreateUserPayment
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IUserPayment'
            description: ""
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: userPayment
                    required: true
                    schema:
                        type: object
    '/api/userPayments/{id}':
        get:
            operationId: UserPaymentsGetUserPayment
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IUserPayment'
            description: ""
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
        put:
            operationId: UserPaymentsUpdateUserPayment
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IUserPayment'
            description: ""
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
                -
                    description: ""
                    in: body
                    name: userPayment
                    required: true
                    schema:
                        type: object
        delete:
            operationId: UserPaymentsDeleteUserPayment
            produces:
                - text/html
            responses:
                '204':
                    description: 'No content'
            description: ""
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
    /api/userRequests:
        get:
            operationId: UserRequestsGetUserRequests
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/IUserRequest'
            description: ""
            parameters: []
        post:
            operationId: UserRequestsCreateUserRequest
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IUserRequest'
            description: ""
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: userRequest
                    required: true
                    schema:
                        type: object
    '/api/userRequests/{id}':
        get:
            operationId: UserRequestsGetUserRequest
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IUserRequest'
            description: ""
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
        put:
            operationId: UserRequestsUpdateUserRequest
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IUserRequest'
            description: ""
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
                -
                    description: ""
                    in: body
                    name: userRequest
                    required: true
                    schema:
                        type: object
        delete:
            operationId: UserRequestsDeleteUserRequest
            produces:
                - text/html
            responses:
                '204':
                    description: 'No content'
            description: ""
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
    /api/userSettings:
        get:
            operationId: UserSettingsGetUserSettings
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/IUserSetting'
            description: ""
            parameters: []
        post:
            operationId: UserSettingsCreateUserSetting
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IUserSetting'
            description: ""
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: userSetting
                    required: true
                    schema:
                        type: object
    '/api/userSettings/{id}':
        get:
            operationId: UserSettingsGetUserSetting
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IUserSetting'
            description: ""
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
        put:
            operationId: UserSettingsUpdateUserSetting
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IUserSetting'
            description: ""
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
                -
                    description: ""
                    in: body
                    name: userSetting
                    required: true
                    schema:
                        type: object
        delete:
            operationId: UserSettingsDeleteUserSetting
            produces:
                - text/html
            responses:
                '204':
                    description: 'No content'
            description: ""
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
    /api/userSpecBets:
        get:
            operationId: UserSpecialBetsGetUserSpecialBets
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/IUserSpecialBet'
            description: ""
            parameters: []
        post:
            operationId: UserSpecialBetsCreateUserSpecialBet
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IUserSpecialBet'
            description: ""
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: userSpecialBet
                    required: true
                    schema:
                        type: object
    '/api/userSpecBets/{id}':
        get:
            operationId: UserSpecialBetsGetUserSpecialBet
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IUserSpecialBet'
            description: ""
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
        put:
            operationId: UserSpecialBetsUpdateUserSpecialBet
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IUserSpecialBet'
            description: ""
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
                -
                    description: ""
                    in: body
                    name: userSpecialBet
                    required: true
                    schema:
                        type: object
        delete:
            operationId: UserSpecialBetsDeleteUserSpecialBet
            produces:
                - text/html
            responses:
                '204':
                    description: 'No content'
            description: ""
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
swagger: '2.0'
securityDefinitions: {}